create chaining for paxos decide phase

+++ src/deptran/paxos/coordinator.h
@@ -29,7 +29,7 @@ class CoordinatorMultiPaxos : public Coordinator {
                         ClientControlServiceImpl *ccsi,
                         uint32_t thread_id);
   ballot_t curr_ballot_ = 1; // TODO
-  uint32_t n_replica_ = 0;   // TODO
+  uint64_t n_replica_ = 0;   // TODO
   slotid_t slot_id_ = 0;
   slotid_t *slot_hint_ = nullptr;
